+++ backend/src/config.py
@@ -10,7 +10,9 @@ class Settings(BaseSettings):
     ALGORITHM: str = "HS256"
     ACCESS_TOKEN_EXPIRE_MINUTES: int = 30
     OPENAI_API_KEY: str = f"{os.getenv('OPENAI_API_KEY')}"
-    CLAUDE_API_KEY: str = f"{os.getenv('CLAUDE_API_KEY')}"
+    OPENAI_MODEL: str = f"{os.getenv('OPENAI_MODEL')}"
+    ANTHROPIC_API_KEY: str = f"{os.getenv('ANTHROPIC_API_KEY')}"
+    ANTHROPIC_MODEL: str = f"{os.getenv('ANTHROPIC_MODEL')}"
     SENTRY_DSN: str = f"{os.getenv('SENTRY_DSN')}"
     CORS_ORIGINS: list = ["http://localhost:8080"]
