+++ docker-compose.yml
@@ -20,7 +20,9 @@ services:
       - MODULE_NAME=app.main
       - VARIABLE_NAME=app
       - OPENAI_API_KEY=${OPENAI_API_KEY}
-      - CLAUDE_API_KEY=${CLAUDE_API_KEY}
+      - OPENAI_MODEL=${OPENAI_MODEL}
+      - ANTHROPIC_API_KEY=${ANTHROPIC_API_KEY}
+      - ANTHROPIC_MODEL=${ANTHROPIC_MODEL}
       - SENTRY_DSN=${SENTRY_DSN}
     command: uvicorn main:app --host 0.0.0.0 --port 8000 --reload
